decouple spec utility functions from misc.py
